fix warning in afl-fuzz-run, tested with '-m32' also

+++ src/afl-fuzz-run.c
@@ -222,7 +222,7 @@ void write_to_testcase(afl_state_t *afl, void *mem, u32 len) {
                                                         len, &new_buf);
 
     if (unlikely(new_size <= 0 || !new_buf))
-      FATAL("Custom_pre_save failed (ret: %ld)", new_size);
+      FATAL("Custom_pre_save failed (ret: %lu)", (long unsigned)new_size);
 
     /* everything as planned. use the new data. */
     ck_write(fd, new_buf, new_size, afl->fsrv.out_file);
